Fix incorrect lines being drawn on the map to old positions
When we received a single position update, we were adding it to the end of the list
(where the oldest position update was), instead of the front.  This would
cause an incorrect path to be drawn from the current position to the last
position in the list.

So now we just add the new update to the front of the list to maintain the
correct order.